Refactored GameFacade and minor fixes
I updated some Java docs to more accurately reflect what methods do. Fixed a few typos as well.
The big changes in this commit are the GameFacade. When a user requests to create a game, it places them in a queue.
Sockets are not yet implemented, but when they are this is how the game cookie and game itself will be sent back to the clients.